👍feat:広告の遷移先を追加
