+++ src/components/layouts/RightBar.tsx
@@ -1,6 +1,7 @@
 import { Box, Button, HStack, Text, VStack } from '@chakra-ui/react'
 import Link from 'next/link'
 import { ReactNode } from 'react'
+import { advertisement } from '../../entities/Advertisement'
 import { SPONSERS } from '../../entities/Sponser'
 import QAButton from '../atoms/QAButton'
 import { SponserBanner } from '../common/suponser/Banner'
@@ -10,12 +11,15 @@ interface Props {
     children?: ReactNode
 }
 
+const sponsers = advertisement.getSome(2)
+
 export const RightBar = ({ children }: Props): JSX.Element => {
     return (
         <VStack as="aside" bgColor="#FBF6F0" h={RIGHT_BAR_HEIGHT} w={RIGHT_BAR_WIDTH} minW={RIGHT_BAR_MINI_WIDTH}>
             <VStack spacing="15px" margin="30px 10px">
-                <SponserBanner sponser={SPONSERS.tasuki} />
-                <SponserBanner sponser={SPONSERS.shinnichi} />
+                {sponsers.map((ad, i) => {
+                    return <SponserBanner sponser={ad} key={i} />
+                })}
             </VStack>
             <QAButton />
         </VStack>
